feat: uint static factory methods
